TLUI を TypeScript に書き換え
