+++ src/content/tlui/components/Component.ts
@@ -1,0 +1,9 @@
+/**
+ * コンポーネント
+ */
+export interface Component {
+    /**
+     * コンポーネントの実要素
+     */
+    element: Element;
+}
